+++ frontend/src/pages/Demo/DemoTailwind.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Typography from "tw-components/Typography";
 import { Button, IconButton } from "tw-components/Buttons";
-import Checkbox from "tw-components/Checkbox";
+import { Checkbox } from "tw-components/Checkbox";
 
 const DemoTailwind = () => {
   // Add a setDarkMode for testing dark mode styles
@@ -15,6 +15,10 @@ const DemoTailwind = () => {
     }
   }, [darkMode]);
 
+  function checkboxOnChange(e: React.ChangeEvent<HTMLInputElement>) {
+    console.log(e.target.checked);
+  }
+
   return (
     <div className="min-h-screen bg-grey-light p-8">
       <div className="mx-auto max-w-4xl rounded-lg bg-white p-6 shadow-md transition-colors duration-300 dark:bg-[#030D2D]">
@@ -23,7 +27,12 @@ const DemoTailwind = () => {
         <Checkbox label="Enabled unselected" />
         <Checkbox label="Disabled selected" defaultChecked disabled />
         <Checkbox label="Disabled unselected" disabled />
-        <Checkbox label="Enabled selected" defaultChecked labelHidden />
+        <Checkbox
+          label="Enabled selected"
+          defaultChecked
+          labelHidden
+          onChange={checkboxOnChange}
+        />
       </div>
       <div className="mx-auto max-w-4xl rounded-lg bg-white p-6 shadow-md transition-colors duration-300 dark:bg-[#030D2D]">
         <Typography.Title1 className="mb-4 text-blue-dark dark:text-white">
